Add module for components and refactor some views to use material3

+++ app/src/main/java/com/example/sleepschedule/ui/dialogs/DialogDeleteEvent.kt
@@ -1,8 +1,13 @@
 package com.example.sleepschedule.ui.dialogs
 
+import android.content.res.Configuration
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.*
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -15,49 +20,54 @@ import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
 import com.example.sleepschedule.R
 import com.example.sleepschedule.common.TimeHelper
+import com.example.sleepschedule.common.scheduleEventMockList
 import com.example.sleepschedule.ui.components.HeaderDialog
 import com.ulises.theme.SleepScheduleTheme
-import com.example.sleepschedule.ui.utils.demoEvent
 import models.ScheduledEvent
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun DialogDeleteEvent(
     event: ScheduledEvent?,
     onDismiss: () -> Unit,
     onDelete: () -> Unit
 ) {
-    Dialog(onDismissRequest = { onDismiss() }) {
-        Card {
+    AlertDialog(
+        onDismissRequest = onDismiss
+    ) {
+        Surface {
             Column(
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.spacedBy(12.dp),
                 modifier = Modifier
                     .fillMaxWidth()
-                    .padding(16.dp)
             ) {
-                HeaderDialog { onDismiss() }
-                Image(
-                    painter = painterResource(id = R.drawable.img_delete),
-                    contentDescription = null,
-                    modifier = Modifier.size(60.dp)
-                )
-                Text(
-                    text = buildAnnotatedString {
-                        append(stringResource(id = R.string.dialog_delete_message))
-                        withStyle(style = SpanStyle(fontWeight = FontWeight.Bold)) {
-                            val time = TimeHelper.convertToHumanReadable(event?.date ?: "")
-                            append(" ${time.split(',')[1]}")
-                        }
-                        append(stringResource(id = R.string.dialog_delete_message_end_message))
-                    },
-                    fontSize = 18.sp,
-                    modifier = Modifier.fillMaxWidth()
-                )
-                Button(onClick = { onDelete() }) {
-                    Text(text = stringResource(id = R.string.dialog_delete_button))
+                HeaderDialog(onClickIcon = onDismiss)
+                Column(
+                    verticalArrangement = Arrangement.spacedBy(12.dp),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                    modifier = Modifier.padding(start = 16.dp, end = 16.dp, bottom = 16.dp)
+                ) {
+                    Image(
+                        painter = painterResource(id = R.drawable.img_delete),
+                        contentDescription = null,
+                        modifier = Modifier.size(60.dp)
+                    )
+                    Text(
+                        text = buildAnnotatedString {
+                            append(stringResource(id = R.string.dialog_delete_message))
+                            withStyle(style = SpanStyle(fontWeight = FontWeight.Bold)) {
+                                val time = TimeHelper.convertToHumanReadable(event?.date ?: "")
+                                append(" ${time.split(',')[1]}")
+                            }
+                            append(stringResource(id = R.string.dialog_delete_message_end_message))
+                        },
+                        fontSize = 18.sp,
+                        modifier = Modifier.fillMaxWidth()
+                    )
+                    Button(onClick = { onDelete() }) {
+                        Text(text = stringResource(id = R.string.dialog_delete_button))
+                    }
                 }
             }
         }
@@ -65,11 +75,12 @@ fun DialogDeleteEvent(
 }
 
 @Preview
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
 fun Prev_DialogDeleteEvent() {
     SleepScheduleTheme {
         DialogDeleteEvent(
-            event = demoEvent,
+            event = scheduleEventMockList[0],
             onDismiss = {},
             onDelete = {}
         )
